+++ common/shared/eslint/index.js
@@ -1,5 +1,6 @@
 exports.baseRules = {
     curly: ['error', 'multi-line'],
+    'eol-last': ['error', 'always'],
     'import/no-cycle': 'error',
     'import/no-self-import': 'error',
     'ts/no-explicit-any': 'warn',
