+++ packages/core/src/common/plugin.ts
