+++ packages/core/src/common/registry.ts
