+++ packages/core/src/shared/__tests__/common.spec.ts
