+++ packages/core/src/shared/__tests__/range.spec.ts
