+++ packages/core/src/shared/__tests__/rectangle.spec.ts
