+++ packages/core/src/shared/__tests__/ref-alias.spec.ts
